Initial Refactor of the entire Code base
