+++ .gitignore
@@ -4,3 +4,4 @@
 venv
 __pycache__/
 .pytest_cache
+.ipynb_checkpoints/
