+++ torch_metrics/__init__.py
@@ -1,13 +1,2 @@
-from torch_metrics.rsquared import RSquared
-from torch_metrics.mse import MeanSquaredError
-from torch_metrics.mae import MeanAbsoluteError
-from torch_metrics.rmse import RootMeanSquaredError
-from torch_metrics.pr import Precision, Recall
-from torch_metrics.f1 import F1Score
-from torch_metrics.meaniou import MeanIoU
-from torch_metrics.dice import DSC
-from torch_metrics.logcosh import LogCoshError
-from torch_metrics.accuracy import Accuracy
-from torch_metrics.kldivergence import KLDivergence
-from torch_metrics.msle import MeanSquaredLogarithmicError
-from torch_metrics.huber import Huber
+from torch_metrics import classification
+from torch_metrics import regression
