+++ torch_metrics/classification/__init__.py
@@ -1,0 +1,7 @@
+from torch_metrics.classification.accuracy import Accuracy
+from torch_metrics.classification.dice import DSC
+from torch_metrics.classification.f1 import F1Score
+from torch_metrics.classification.hinge import HingeMetric
+from torch_metrics.classification.kldivergence import KLDivergence
+from torch_metrics.classification.pr import Precision, Recall
+from torch_metrics.classification.squarehinge import SquareHingeMetric
