+++ torch_metrics/classification/squarehinge.py
