+++ torch_metrics/regression/__init__.py
@@ -1,0 +1,8 @@
+from torch_metrics.regression.huber import Huber
+from torch_metrics.regression.logcosh import LogCoshError
+from torch_metrics.regression.mae import MeanAbsoluteError
+from torch_metrics.regression.meaniou import MeanIoU
+from torch_metrics.regression.mse import MeanSquaredError
+from torch_metrics.regression.msle import MeanSquaredLogarithmicError
+from torch_metrics.regression.rmse import RootMeanSquaredError
+from torch_metrics.regression.rsquared import RSquared
