+++ torch_metrics/regression/logcosh.py
